NewProfile as value object
