+++ packages/io-auth-n-identity-domain/src/domain/value-objects/profile/profile.vo.ts
@@ -1,3 +1,4 @@
+import { EmailAddressSchema } from "@pagopa/hexagonal-core";
 import { z } from "zod";
 
 export const IsEmailValidatedSchema = z.boolean();
@@ -84,3 +85,13 @@ export const PushNotificationsContentTypeSchema = z.enum(
 export type PushNotificationsContentTypeSchema = z.infer<
   typeof PushNotificationsContentTypeSchema
 >;
+
+const NewProfileBase = z.object({
+  is_email_validated: IsEmailValidatedSchema,
+});
+
+export const NewProfile = NewProfileBase.extend({
+  email: EmailAddressSchema.optional(),
+  is_test_profile: IsTestProfileSchema.optional(),
+});
+export type NewProfile = z.infer<typeof NewProfile>;
